fix(client) refactor to useCommitStatus hook instead of component
